+++ src/handlers/updateConfiguration.ts
@@ -3,7 +3,6 @@ import { BASE_TYPES, assertTypeNameIsRestricted } from "../constants";
 import {
   getNeo4jDriver,
   inferRelationshipFieldName,
-  insertion,
   toPlural,
 } from "../utilities";
 import { toGenericStruct, toGraphQLTypeDefs } from "@neo4j/introspector";
@@ -47,6 +46,13 @@ export async function doUpdateConfiguration(
   const collectionNames = Object.values(genericStruct.nodes)
     .map((n) => n.labels[0])
     .map(toPlural);
+
+  for (const label of collectionNames) {
+    if (assertTypeNameIsRestricted(label)) {
+      throw new Error(`${label} is a restricted name!`);
+    }
+  }
+
   configuration.config = {
     collection_names: collectionNames,
     object_types: { ...BASE_TYPES },
@@ -58,40 +64,13 @@ export async function doUpdateConfiguration(
   };
   genericStructToHasuraConfig(genericStruct, configuration.config);
 
-  for (const label of collectionNames) {
-    // TODO: needed?
-    if (assertTypeNameIsRestricted(label)) {
-      throw new Error(`${label} is a restricted name!`);
-    }
-    /*
-    const { records } = await driver.executeQuery(
-      `MATCH (n: ${label}) RETURN n LIMIT 1`
-    );
-
-    if (records.length > 0) {
-      const recordPayload = records[0].get("n")["properties"];
-
-      const fieldDict = insertion(
-        label,
-        recordPayload,
-        configuration.config.object_types
-      );
-      configuration.config.object_types[label] = {
-        description: null,
-        fields: fieldDict,
-      };
-      configuration.config.object_fields[label] = Object.keys(fieldDict);
-    }
-    */
-  }
-
   return configuration;
 }
 
 type NodeInfo = {
   k: string;
   labels: string[];
-  label: string;
+  mainLabel: string;
   properties: Property[];
 };
 
@@ -114,22 +93,22 @@ function genericStructToHasuraConfig(
     acc[k] = {
       k,
       labels: node.labels,
-      label: node.labels[0], // TODO: Assumption - generated query (the collection) will be named using the first label only
+      mainLabel: node.labels[0], // TODO: Assumption - generated query (the collection) will be named using the first label only
       properties: node.properties,
     };
     return acc;
   }, {});
   for (const k in nodesMap) {
     const node = nodesMap[k];
-    const nodeLabel = node.label;
-    console.log("label", nodeLabel);
+    const nodeMainLabel = node.mainLabel;
+    console.log("label", nodeMainLabel);
     const fields = propertiesToHasuraField(node.properties);
     console.log("fields", fields);
-    config.object_types[nodeLabel] = {
+    config.object_types[nodeMainLabel] = {
       description: null,
       fields: Object.fromEntries(fields.entries()),
     };
-    config.object_fields[nodeLabel] = Array.from(fields.keys());
+    config.object_fields[nodeMainLabel] = Array.from(fields.keys());
   }
 
   const relationships = Object.entries(
@@ -148,8 +127,8 @@ function genericStructToHasuraConfig(
             type: rel.type,
             fieldName: inferRelationshipFieldName(
               rel.type,
-              fromType?.label as string,
-              toType?.label as string,
+              fromType?.mainLabel as string,
+              toType?.mainLabel as string,
               "OUT"
             ), // typenames can differ from what introspector generated in type defs
           },
@@ -160,8 +139,8 @@ function genericStructToHasuraConfig(
             type: rel.type,
             fieldName: inferRelationshipFieldName(
               rel.type,
-              fromType?.label as string,
-              toType?.label as string,
+              fromType?.mainLabel as string,
+              toType?.mainLabel as string,
               "IN"
             ), // typenames can differ from what introspector generated in type defs
           },
@@ -176,13 +155,13 @@ function genericStructToHasuraConfig(
     }
 
     // TODO: keys should be fields with Unique constraint if it exists
-    const fromTypeKey = config.object_fields[rel.fromType.label][0];
-    const toTypeKey = config.object_fields[rel.toType.label][0];
+    const fromTypeKey = config.object_fields[rel.fromType.mainLabel][0];
+    const toTypeKey = config.object_fields[rel.toType.mainLabel][0];
 
-    config.foreign_keys[rel.fromType.label] = {
-      ...config.foreign_keys[rel.fromType.label],
+    config.foreign_keys[rel.fromType.mainLabel] = {
+      ...config.foreign_keys[rel.fromType.mainLabel],
       [rel.fieldName]: {
-        foreign_collection: toPlural(rel.toType.label),
+        foreign_collection: toPlural(rel.toType.mainLabel),
         column_mapping: { [fromTypeKey]: toTypeKey },
       },
     };
